Fix NaNs in positive fraction
Days where the new case count is zero or where test values
aren't reported lead to inf or NaN values for positiveFraction.
This tends to make the plots ugly and screws up 7 day averages.

As a workaround, the code now replaces NaN and inf values with the
previous valid value of positiveFraction.