Move away from AWS Lambda
The latest deploy hit a Lambda size restriction.
It's too much hassle to figure out how to keep the python
libraries under the size limit and I'm too cheap to pay for
an EC2 instance, so I'm planning to run under Apache on a
raspberry pi W.

This commit is the first step in that direction.