Mods to run flask app on Rasperry Pi Zero W
I was unable to deploy due to AWS Lambda resource limits
(numpy and other modules exceed the max size for a Lambda).

I tried switching to google cloud, but quit when it looked
like the cost was going to be nonzero.

I could go back to running on an EC2 instance, but I already have
an idle pi zero W, so why pay for EC2?

The static files are still served from S3.  The pi zero w takes
awhile to run the update process, but that doesn't affect the
end users so who cares?