Add fix for Turkey
The Turkish new case data for December 10, 2020 was affected
by a correction to overall case numbers that Turkey applied
on that date.  This fix makes the plot choose a more reasonable
scale by setting the December 10th new case number to zero.

It's not a perfect choice, but the plot is more readable.